Tried fixing circular import
